feat: changes that need to meet our RSG Agreement req
Per the agreement, we require that you use the banners we've created for
your event here. I see Scrum Alliance displayed on your website - are we
listed as the title sponsor? Please update so that our logo links to our
webpage https://www.scrumalliance.org/,  We ask that you add a tab to
the webpage that says Scrum Alliance and also links to our webpage. A
great example of an RSG that meets these website requirements is RSG
Cape Town - you can review their website here.